requirements.txt
torch>=2.0.0
transformers>=4.30.0
gradio>=3.50.0
peft>=0.5.0
accelerate>=0.21.0
bitsandbytes>=0.41.0
datasets>=2.14.0
boto3>=1.28.0
kubernetes>=28.1.0
psutil>=5.9.0
wandb>=0.15.0
mlflow>=2.8.0
gputil>=1.4.0
deepspeed>=0.10.0
safetensors>=0.4.0
tensorboard>=2.14.0
paramiko>=3.4.0
modelscope>=1.9.0
fastapi>=0.104.0
pydantic>=2.4.2
PyYAML>=6.0.1
python-multipart>=0.0.6
uvicorn>=0.23.2
